[feature/ProblemUpdate] fix: rename job class name and BatchController to QuartzService
