+++ .gitignore
@@ -25,7 +25,7 @@ bin/
 out/
 !**/src/main/**/out/
 !**/src/test/**/out/
-!**/src/main/generated/
+!**/src/main/generated/**
 
 ### NetBeans ###
 /nbproject/private/
