+++ src/main/java/cos/peerna/domain/Category.java
@@ -3,7 +3,7 @@ package cos.peerna.domain;
 import com.fasterxml.jackson.annotation.JsonCreator;
 
 public enum Category {
-    JAVA, JAVASCRIPT, PYTHON, SPRING, REACT, NODE_JS, DJANGO, RDB, NOSQL, OS, SECURITY, NETWORK, DATA_STRUCTURE, ALGORITHM, COMPILER, TROUBLE, DESIGN, TEST, INFRA, CONTAINER, DEVOPS, COMMUNICATION, WORK, TECH, CODING_TEST;
+    JAVA, JAVASCRIPT, PYTHON, SPRING, REACT, NODEJS, DJANGO, RDB, NOSQL, OS, SECURITY, NETWORK, DATA_STRUCTURE, ALGORITHM, COMPILER, TROUBLE, DESIGN, TEST, INFRA, CONTAINER, DEVOPS, COMMUNICATION, WORK, TECH, CODING_TEST;
 
     @JsonCreator
     public static Category from(String s) {
